added api calls to api dir and made it so ui is updated as fields are updated
