+++ frontend/src/api/user.ts
@@ -1,10 +1,34 @@
 import api from "./index";
+import {isAxiosError} from "axios";
+import {UserProfile} from "@/pages/Account/Account.tsx";
 
 async function getUser() {
-  return api.get('/me')
-    .then(response => {return response.data});
+    return api.get('/me')
+        .then(response => {
+            return response.data
+        });
+}
+
+
+async function updateUser(userId: number, data: UserProfile) {
+    try {
+        const updated_user = await api.put(`/api/users/${userId}`, data);
+        console.log(updated_user.data);
+        return updated_user.data;
+    } catch (error) {
+        if (isAxiosError(error)) {
+            console.error('Error updating user:', {
+                status: error.response?.status,
+                data: error.response?.data,
+            });
+        } else {
+            console.error('Unexpected error:', error);
+        }
+        throw error;
+    }
 }
 
 export {
-  getUser
+    getUser,
+    updateUser
 }
